Rename file to mrt_overseer.lua
